refactor: create converters module
